requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
contourpy==1.3.2
cycler==0.12.1
Django==5.2.12
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
fonttools==4.61.1
idna==3.11
inflection==0.5.1
kiwisolver==1.4.9
matplotlib==3.10.8
mysqlclient==2.2.8
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
requests==2.32.5
scipy==1.15.3
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
